pytest: mbox multi sys: check only success message
The other two messages get output concurrently with another
message ("reply_rcv") that comes from a concurrent task and
ends up in the middle of the message we are looking for,
so we fail to match.

Hopefully the 'success' message is not subject to this overlap,
because it happens a bit later (after the concurrent task
should be already done with its messages).