pytest: check TRCH boot
We need to consume the output or else the Qemu thread for TRCH gets
halted when the FIFO fills up. This one epect here is only a partial
solution: it consumes part of the output (most), but nothing after that.
Full solution comming in a later commit.

NOTE: this breaks the MMU tests because they look through TRCH
output. Fix comming soon.