pytest: thread per pexpect connection
Background threads for consuming the output (added earlier) are needed
to get the full log. But, the existing design required passing
pexpect handles across threads, which seemed to work, but is not
supported according to documentation. Since this was suspected as
the source of problems, the present patch changes the design to
keep the handles in the thread that called spawn.