pytest: remove pyserial middleman
Open and hand the file descriptor to theh PTY slave device to pexpect
directly. I think this might be the patch that fixed stalls.

Background: Canonincal use case of pexpect is the where pexpect
is given a path to an executable, pexpect creates a PTY master,
and a PTY slave, executes the binary, and attaches the binary's
stdin/stdout to the PTY slave; pexpect then uses the master
to interact with the binary (exactly the way the human user's terminal
would be interacting with the binary launched by a shell). In
our case, we already have an existing PTY slave (allocated by Qemu) with
an "application" attached to it (the machine model), and all
we want pexpect to do is a subset of what it would normally do:
we only want it to connect to this PTY slave.

PySerial is also capable of attaching to PTY slaves (a serial port
is a kind of like a PTY slave, I think). So, we could use either
PySerial or Pexpect to interact with the "serial" port. I am
not sure what exactly does giving PySerial's file object to Pexpect
result in (does Pexpect simply call .fileno(), and use os read/write
syscalls on the fd directly as opposed to file object's read/write
methods? if so, does PySerial return the file descriptor for the
underlying device?  if so, then the effect of chaining amounts to
PySerial opening the slave with particular flags and setting it
up with particular ioctls which may or may not be the right ones;
if not, then the situation might be much worse where data actually
flows through PySerial and pexpect effectively gets a dumb pipe
which is very different from a PTY slave (because ioctl)). But,
whatever that does, let's just not do it.